test: 예약 대기 기능 테스트 추가 및 rank 타입 정합성 수정

+++ src/main/java/roomescape/domain/waiting/service/ReserveWaitingService.java
@@ -63,7 +63,7 @@ public class ReserveWaitingService {
         }
 
         ReserveWaiting newReserveWaiting = reserveWaitingRepository.save(new ReserveWaiting(foundMember, date, foundTime, foundTheme));
-        int aheadCount = reserveWaitingRepository.countByDateAndTimeAndThemeAndIdLessThan(date, foundTime, foundTheme, newReserveWaiting.getId());
+        Long aheadCount = reserveWaitingRepository.countByDateAndTimeAndThemeAndIdLessThan(date, foundTime, foundTheme, newReserveWaiting.getId());
 
         return new WaitingWithRank(newReserveWaiting, aheadCount);
     }
