+++ src/main/java/roomescape/domain/waiting/web/dto/WaitingResponse.java
@@ -4,7 +4,7 @@ import roomescape.domain.waiting.service.result.WaitingWithRank;
 
 public record WaitingResponse(
         Long id,
-        Integer waitingNumber
+        Long waitingNumber
 ) {
     public static WaitingResponse from(WaitingWithRank waitingWithRank) {
         return new WaitingResponse(waitingWithRank.reserveWaiting().getId(), waitingWithRank.rank());
